+++ libs/windy-sounding/src/Plugin.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import { onDestroy, onMount } from 'svelte';
   import { openPlugin, mountPlugin, destroyPlugin } from './sounding';
+  import { loadSetting, Settings } from './util/settings';
 
   let pluginElement: HTMLElement;
 
@@ -13,10 +14,23 @@
       [parameters.lat, parameters.lon, parameters.modelName] = [parameters.modelName, parameters.lat, W.store.get('product')];
     }
 
+    let lat = parameters?.lat;
+    let lon = parameters?.lon;
+
+    if (lat === undefined || lon === undefined) {
+      try {
+        const location = JSON.parse(loadSetting(Settings.location));
+        lat = location.lat;
+        lon = location.lon;
+      } catch {
+        // empty
+      }
+    }
+
     const mapCenter = W.map.map.getCenter();
-    const lat = Number(parameters?.lat ?? mapCenter.lat);
-    const lon = Number(parameters?.lon ?? mapCenter.lng);
-    const modelName = parameters?.modelName ?? W.store.get('product');
+    lat = Number(lat ?? mapCenter.lat);
+    lon = Number(lon ?? mapCenter.lng);
+    const modelName = parameters?.modelName ?? loadSetting(Settings.model) ?? W.store.get('product');
     openPlugin({ lat, lon, modelName });
   };